[jscodeshift] Split into multiple files
