+++ types/jscodeshift/tslint.json
@@ -1,8 +1,3 @@
 {
-    "extends": "dtslint/dt.json",
-    "rules": {
-        "no-declare-current-package": false,
-        "no-self-import": false,
-        "no-unnecessary-generics": false
-    }
+    "extends": "dtslint/dt.json"
 }
